requirements.txt
streamlit
telethon
aiosqlite
python-dotenv
PySocks
streamlit-autorefresh